feat: Custom CSS now also applies to settings window
